fix: 더미데이터파일만듦, 제목도 데티어틀 받아오게변경(#22)
